@@ -3,25 +3,6 @@
   "proxy": "http://127.0.0.1:5000",
   "version": "0.1.0",
   "private": true,
-  "dependencies": {
-    "@react-google-maps/api": "^2.7.0",
-    "@testing-library/jest-dom": "^5.15.0",
-    "@testing-library/react": "^11.2.7",
-    "@testing-library/user-event": "^12.8.3",
-    "react": "^17.0.2",
-    "react-axios": "^2.0.5",
-    "react-dom": "^17.0.2",
-    "react-google-charts": "^4.0.0",
-    "react-paypal-button-v2": "^2.6.3",
-    "react-redux": "^7.2.6",
-    "react-responsive-carousel": "^3.2.23",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.1.2",
-    "redux-thunk": "^2.4.0",
-    "socket.io-client": "^4.4.1",
-    "web-vitals": "^1.1.2"
-  },
   "scripts": {
     "start": "react-scripts start",
     "build": "react-scripts build",
@@ -45,5 +26,25 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "dependencies": {
+    "@react-google-maps/api": "^2.10.2",
+    "@testing-library/jest-dom": "^5.16.4",
+    "@testing-library/react": "^13.1.1",
+    "@testing-library/user-event": "^14.1.1",
+    "axios": "^0.27.2",
+    "react": "^18.1.0",
+    "react-axios": "^2.0.5",
+    "react-dom": "^18.1.0",
+    "react-google-charts": "^4.0.0",
+    "react-paypal-button-v2": "^2.6.3",
+    "react-redux": "^8.0.1",
+    "react-responsive-carousel": "^3.2.23",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
+    "redux": "^4.2.0",
+    "redux-thunk": "^2.4.1",
+    "socket.io-client": "^4.5.0",
+    "web-vitals": "^2.1.4"
   }
 }
