@@ -1,6 +1,7 @@
 import React, { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link, useParams } from "react-router-dom";
+import { useNavigate } from "../../node_modules/react-router-dom/index";
 import { listProducts } from "../actions/productActions";
 import LoadingBox from "../components/LoadingBox";
 import MessageBox from "../components/MessageBox";
@@ -8,7 +9,8 @@ import Product from "../components/Product";
 import Rating from "../components/Rating";
 import { prices, ratings } from "../utils";
 
-export default function SearchScreen(props) {
+export default function SearchScreen() {
+  const navigate = useNavigate();
   const {
     name = "all",
     category = "all",
@@ -67,7 +69,7 @@ export default function SearchScreen(props) {
           <select
             value={order}
             onChange={(e) => {
-              props.history.push(getFilterUrl({ order: e.target.value }));
+              navigate(getFilterUrl({ order: e.target.value }));
             }}
           >
             <option value="newest">Newest Arrivals</option>
