@@ -1,116 +1,18 @@
 import React, { useEffect, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
+import {
+  useNavigate,
+  useParams,
+} from "../../node_modules/react-router-dom/index";
 import { detailsUser, updateUser } from "../actions/userActions";
 import LoadingBox from "../components/LoadingBox";
 import MessageBox from "../components/MessageBox";
 import { USER_UPDATE_RESET } from "../constants/orderConstants";
 
-// export default function UserEditScreen(props) {
-//   const userId = props.match.params.id;
-//   const [name, setName] = useState("");
-//   const [email, setEmail] = useState("");
-//   const [isSeller, setIsSeller] = useState(false);
-//   const [isAdmin, setIsAdmin] = useState(false);
-
-//   const userDetails = useSelector((state) => state.userDetails);
-//   const { loading, error, user } = userDetails;
-
-//   const userUpdate = useSelector((state) => state.userUpdate);
-//   const {
-//     loading: loadingUpdate,
-//     error: errorUpdate,
-//     success: successUpdate,
-//   } = userUpdate;
-
-//   const dispatch = useDispatch();
-//   useEffect(() => {
-//     if (successUpdate) {
-//       dispatch({ type: USER_UPDATE_RESET });
-//       props.history.push("/userlist");
-//     }
-//     if (!user) {
-//       dispatch(detailsUser(userId));
-//     } else {
-//       setName(user.name);
-//       setEmail(user.email);
-//       setIsSeller(user.isSeller);
-//       setIsAdmin(user.isAdmin);
-//     }
-//   }, [dispatch, props.history, successUpdate, user, userId]);
-
-//   const submitHandler = (e) => {
-//     e.preventDefault();
-//     // dispatch update user
-//     dispatch(updateUser({ _id: userId, name, email, isSeller, isAdmin }));
-//   };
-//   return (
-//     <div>
-//       <form className="form" onSubmit={submitHandler}>
-//         <div>
-//           <h1>Edit User {name}</h1>
-//           {loadingUpdate && <LoadingBox></LoadingBox>}
-//           {errorUpdate && (
-//             <MessageBox variant="danger">{errorUpdate}</MessageBox>
-//           )}
-//         </div>
-//         {loading ? (
-//           <LoadingBox />
-//         ) : error ? (
-//           <MessageBox variant="danger">{error}</MessageBox>
-//         ) : (
-//           <>
-//             <div>
-//               <label htmlFor="name">Name</label>
-//               <input
-//                 id="name"
-//                 type="text"
-//                 placeholder="Enter name"
-//                 value={name}
-//                 onChange={(e) => setName(e.target.value)}
-//               ></input>
-//             </div>
-//             <div>
-//               <label htmlFor="email">Email</label>
-//               <input
-//                 id="email"
-//                 type="email"
-//                 placeholder="Enter email"
-//                 value={email}
-//                 onChange={(e) => setEmail(e.target.value)}
-//               ></input>
-//             </div>
-//             <div>
-//               <label htmlFor="isSeller">Is Seller</label>
-//               <input
-//                 id="isSeller"
-//                 type="checkbox"
-//                 checked={isSeller}
-//                 onChange={(e) => setIsSeller(e.target.checked)}
-//               ></input>
-//             </div>
-//             <div>
-//               <label htmlFor="isAdmin">Is Admin</label>
-//               <input
-//                 id="isAdmin"
-//                 type="checkbox"
-//                 checked={isAdmin}
-//                 onChange={(e) => setIsAdmin(e.target.checked)}
-//               ></input>
-//             </div>
-//             <div>
-//               <button type="submit" className="primary">
-//                 Update
-//               </button>
-//             </div>
-//           </>
-//         )}
-//       </form>
-//     </div>
-//   );
-// }
-
 export default function UserEditScreen(props) {
-  const userId = props.match.params.id;
+  const navigate = useNavigate();
+  const params = useParams();
+  const { id: userId } = params;
   const [name, setName] = useState("");
   const [email, setEmail] = useState("");
   const [isSeller, setIsSeller] = useState(false);
@@ -130,7 +32,7 @@ export default function UserEditScreen(props) {
   useEffect(() => {
     if (successUpdate) {
       dispatch({ type: USER_UPDATE_RESET });
-      props.history.push("/userlist");
+      navigate("/userlist");
     }
     if (!user) {
       dispatch(detailsUser(userId));
@@ -140,7 +42,7 @@ export default function UserEditScreen(props) {
       setIsSeller(user.isSeller);
       setIsAdmin(user.isAdmin);
     }
-  }, [dispatch, props.history, successUpdate, user, userId]);
+  }, [dispatch, navigate, successUpdate, user, userId]);
 
   const submitHandler = (e) => {
     e.preventDefault();
